Added the comment.java, and also implemented Firebase thanks for the help Jonathan and need to push
