+++ QuiRky/app/src/main/java/com/example/quirky/Comment.java
@@ -1,5 +1,39 @@
+// Author: Raymart
+// Contact me Through discord for any questions
+
 package com.example.quirky;
 
-// This class will keep all the attributes of A comment
+import java.util.Date;
+
+// This will add the comment to the firebase Database.
 public class Comment {
+    private String content, uname;
+    private Date timestamp;
+
+    public Comment(String content, String uname, Date timestamp) {
+        this.content = content;
+        this.uname = uname;
+        this.timestamp = timestamp;
+
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public String getUname() {
+        return uname;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public Date getTimestamp() {
+        return timestamp;
+    }
+
+    public void setUname(String uname) {
+        this.uname = uname;
+    }
 }
