+++ QuiRky/app/src/main/java/com/example/quirky/CommentActivity.java
@@ -11,31 +11,46 @@ import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Date;
 import java.util.Locale;
+import java.util.Map;
 
 public class CommentActivity extends AppCompatActivity {
     private Button Save;
     private Button Cancel;
 
     ListView commentList;
-    ArrayList<AddComment> commentDataList;
+    ArrayList<Comment> commentDataList;
     CommentList commentAdapter;
 
-    FirebaseDatabase firebaseDatabase; // Suspect Code will need to review how the Firebase database works.
+//    FirebaseDatabase firebaseDatabase; // Suspect Code will need to review how the Firebase database works.
+
+    DatabaseManager DM;
 
     // Need to retrieve the Comments that were made for that QR code.
     // Also still need to show the comments.
+
+    // A message which contains the QR Code Information from the previous intent wll need to be passed. And then the
+    // The QR Code information will be used to open the specific collection. So that the comment for that QR code will be
+    // read. And be displayed in the list.
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        firebaseDatabase = FirebaseDatabase.getInstance(); // Need to review how the firebase database works.
+//        firebaseDatabase = FirebaseDatabase.getInstance(); // Need to review how the firebase database works.
+
+        String comment = "whdfsk"; // sample comment
+        Map<String, String> a; // The comment must be stored with a random key value pair, but doesn't really matter what the name is
+        a.put("comment", comment); // Key matches the Field that you want to store it in.
+
+        DM.setCollection("comments"); // Opens the collection Comments
+
+        DM.write(a, "comment1"); // Writes a into the document "comment1"
 
         // Need to extract comment values from database.
 
@@ -45,8 +60,9 @@ public class CommentActivity extends AppCompatActivity {
         String[] sampleComments = {"cool", "nice", "rad"};
         String[] sampleNames = {"Akhmetov", "GoldenBear", "Panda"};
 
+
         for (int i = 0; i < sampleComments.length; i++) {
-            commentDataList.add(new AddComment(sampleComments[i], sampleNames[i]));
+            commentDataList.add(new Comment(sampleComments[i], sampleNames[i], new Date()));
         }
 
         commentAdapter = new CommentList(this, commentDataList);
@@ -91,24 +107,25 @@ public class CommentActivity extends AppCompatActivity {
         String content = editTextComment.getText().toString();
 
         // this should be stored in the firestore
-        AddComment comment = new AddComment(content, uName);
+        Comment comment = new Comment(content, uName, new Date());
 
         // Suspect Code
-        DatabaseReference commentReference = firebaseDatabase.getReference("Comment").child(PostKey).push();
-        commentReference.setValue(comment).addOnSuccessListener(new OnSuccessListener<Void>() {
-            @Override
-            public void onSuccess(Void aVoid) {
-                showMessage("Comment Added");
-                editTextComment.setText("");
-            }
-        }).addOnFailtureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception e) {
-                showMessage("fail to add comment : "+e.getMessage());
-            }
-        });
-        finish();
-    }
+//        DatabaseReference commentReference = firebaseDatabase.getReference("Comment").child(PostKey).push();
+
+//        commentReference.setValue(comment).addOnSuccessListener(new OnSuccessListener<Void>() {
+//            @Override
+//            public void onSuccess(Void aVoid) {
+//                showMessage("Comment Added");
+//                editTextComment.setText("");
+//            }
+//        }).addOnFailtureListener(new OnFailureListener() {
+//            @Override
+//            public void onFailure(@NonNull Exception e) {
+//                showMessage("fail to add comment : "+e.getMessage());
+//            }
+//        });
+//        finish();
+//    }
 
     private void showMessage(String message) {
         Toast.makeText(this, message, Toast.LENGTH_LONG).show();
